qontract_api/pyproject.toml
[project]
name = "qontract-api"
version = "0.1.0"
description = "REST API service for qontract reconciliation"
authors = [{ name = "Red Hat App-SRE Team", email = "sd-app-sre@redhat.com" }]
license = { text = "Apache 2.0" }
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.121.1",
    "uvicorn[standard]==0.38.0",
    "celery==5.5.3",
    "valkey==6.1.1",
    "pydantic==2.12.4",
    "pydantic-settings==2.12.0",
    "python-jose[cryptography]==3.5.0",
]


[dependency-groups]
dev = [
    # development and testing tools
    "mypy==1.18.2",
    "pytest==9.0.1",
    "pytest-asyncio==1.3.0",
    "pytest-cov==7.0.0",
    "ruff==0.14.4",
]

[tool.uv.sources]
qontract-reconcile = { workspace = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["qontract_api"]

[tool.ruff]
line-length = 88
target-version = 'py312'
src = ["qontract_api"]
fix = true

[tool.ruff.format]
preview = true

[tool.ruff.lint]
preview = true
select = ["ALL"]
ignore = [
    "ANN401",  # allow typings.Any
    "CPY",     # Missing copyright header
    "D100",    # Missing docstring in public module
    "D101",    # Missing docstring in public class
    "D102",    # Missing docstring in public method
    "D103",    # Missing docstring in public function
    "D104",    # Missing docstring in public package
    "D105",    # Missing docstring in magic method
    "D106",    # Missing docstring in nested public class
    "D107",    # Missing docstring in __init__
    "D203",    # 1 blank line required before class docstring
    "D211",    # No blank lines allowed before class docstring
    "D212",    # multi-line-summary-first-line
    "D213",    # multi-line-summary-second-line
    "D4",      # Doc string style
    "E501",    # Line too long
    "FIX002",  # TODO comment found
    "G004",    # Logging statement uses f-string
    "PLR0904", # Too many public methods
    "PLR0913", # Too many arguments
    "PLR0917", # Too many positional arguments
    "S101",    # Use of assert detected. Pytest uses assert
    "S404",    # subprocess import
    "EM101",   # Exception must not use a string literal, assign to variable first
    "EM102",   # Exception must not use an f-string literal, assign to variable first
    "TRY003",  # Avoid specifying long messages outside the exception class
    "TRY300",  # try-consider-else
    "TC002",   # type-checking-block
    "TC003",   # type-checking-block
    # pydoclint
    "DOC",
    # May conflict with the formatter, https://docs.astral.sh/ruff/formatter/#conflicting-lint-rules
    "W191",
    "E111",
    "E114",
    "E117",
    "D206",
    "D300",
    "Q",
    "COM812",
    "COM819",
    "ISC001",
    # remove me
    "TD", # TODO found
]

[tool.ruff.lint.isort]
known-first-party = ["qontract_api"]

[tool.ruff.lint.pep8-naming]
classmethod-decorators = ["classmethod"]

[[tool.mypy.overrides]]
# Below are all of the packages that don't implement stub packages. Mypy will throw an error if we don't ignore the
# missing imports. See: https://mypy.readthedocs.io/en/stable/running_mypy.html#missing-imports
module = ["celery.*", "jose"]
ignore_missing_imports = true
